// arcadia.backoffice.api/src/modules/history/enums/history.entities.ts
export enum HistoryEntities {
  GROUPS = 'groups',
  MACHINES = 'machines',
  CAMERAS = 'cameras',
  VOUCHERS = 'vouchers',
  OPERATORS = 'operators',
  ALERTS = 'alerts',
  DISPUTES = 'disputes',
  MAINTENANCE = 'maintenance',
  MONITORING = 'monitoring',
  ADMINISTRATION = 'administration',
  PLAYERS = 'players',
  SESSIONS = 'sessions',
}